watdahel

+++ hypergraph.hpp
@@ -6,7 +6,7 @@ typedef struct {
 } hypergraph_edge_t;
 
 typedef struct {
-  uint64_t dist;
+  int32_t dist;
   uint64_t effort;
   uint64_t source;
   uint64_t drain;
